clean unmanaged chezmoi configs but asdf.fish

+++ .chezmoiscripts/run_9_update_fish.sh
@@ -1,10 +1,10 @@
 #!/usr/bin/fish
 
-chezmoi add ~/.config/fish/conf.d/asdf.fish
-
 # cleanup unmanaged files
 for f in $(cz unmanaged ~/.config/fish/conf.d)
-    rm $f
+    if [ "$(basename $f)" != "asdf.fish" ]
+        rm $f
+    end
 end
 
 # update plugins
